requirements.txt
absl-py==0.13.0
aniso8601==9.0.1
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
cycler==0.10.0
decorator==5.0.9
feather==0.1.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
numba==0.53.1
numpy==1.20.2
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
Pillow==8.2.0
pooch==1.4.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
SoundFile==0.10.3.post1
split-folders==0.4.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.2.0
tqdm==4.62.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
